Add Skein MAC to SRTP and enable it in ZRTP. The Skein 32 and Skein 64 MAC authentication is now usable. Refer to ZRTP specification chapter 5 and it subchapters. Skein requires less CPU power to compute a MAC while at the same time is more secure than SHA1 based HMAC.

+++ zsrtp/include/CryptoContext.h
@@ -383,12 +383,12 @@ private:
     int32_t  n_s;
     uint8_t* k_s;
 
-    uint8_t ealg;
-    uint8_t aalg;
-    uint8_t ekeyl;
-    uint8_t akeyl;
-    uint8_t skeyl;
-    uint8_t tagLength;
+    int32_t ealg;
+    int32_t aalg;
+    int32_t ekeyl;
+    int32_t akeyl;
+    int32_t skeyl;
+    int32_t tagLength;
     bool  seqNumSet;
 
     void*   macCtx;
